Adjust column to account for '\r\n' line endings
Editors often won't show but a single line ending. Adjust column for
these lines when column falls on end of line.  This is so we don't
accidentally show the next line.  Some editors will allow showing line
endings, in this case, if col is on \n or \r\n, it will be shown as \r.
It is a small price to pay for better line/col results.