Fix for non OSX platforms and image sizes
Retina support worked great for retina and non-retina on OSX, but it
appears this is because of how OSX handles images; windows and linux
sizing was broken.

Now we conditionally size or scale depending on the platform.  In the
future, it would be quicker to already have normal **and** 2X size image
variants opposed to scale on the fly, but this is a fine work around for
now.