Add back in kill method in rumcore to cleanly abort in loops
Since everything is done with iterators and generators in rumcore, when
iterating the records of rumcore it probably isn’t good to just break
as things in iterator (open files etc.) may not get cleaned up proper.
So instead, we will add back in a kill routine which can be called from
the loop that is iterating the records.  The kill should be propagated
to all the workes (file searcher, dirwalker, etc.).  That way we should
clean up proper.