Revert WxFormbuilder gen for 3.5-RC1
Until I can figure out what is screwing up the alignment when
generating code in the new 3.5-RC1, I need to use an older version.

There seems to be two big changes that may or may not be responsible.

1. Instead of getting the static box from the sizer for parent,
generated code now feeds in the related panel.

2. When adding an element to the sizer, proportion 1 is being used
instead of 0.

These are the only things I see that might be responsible.  I suspect
the sizer proportion is the culprit.  Maybe I will just need to go
through and set proportion to 0 for every sizer.   It seems older
versions ignored size 1 and just used 0.