Fix OSX only recognizing one arg set sent via pipes
In OSX, once one set of arguments has been sent, no more arguments can
be accepting via pipes.  More than likely an EOF is received and the
file can no longer read lines, but I am still not 100% sure what is
going.  A workaround is exit reading the file once a set of arguments
is processed.  Afterwards, we reopen the file and we will be able to
receive more arguments.  Though the reason is only guessed at, it is
clear that something happens to the file stream once the first set of
arguments is read.  This should resolve issue #90.