Use gridbagsizer to fix mac out of order tabbing
FlexGridSizer has a quirk on mac where it will evaluate the first
column first (as expected), but then on the last column it will wrap
the rows back to the first column.  Ugh.

GridBagSizer fixes this as you can align all the columns controls
inline side by side and then when you tab through them on mac, you
process each row in a more intuitive fashion.