Rumcore doesn't decode strings directly
Rumcore will no longer decode strings to simplify the internal code.
If someone is using the API, they should decode it to Unicode or pass
it as a byte string for binary.  Only files will be decoded/encoded on
input/output. Technically everything is included so a user could detect
encoding.  But as the API is not documented right now, I won’t worry
about it, but in the future, maybe we can expose an easy detect
function that does everything the that happens for files.