Tweak date and time control
On each OS, the date and time object is a little different with how it
fits.  Try and ensure the date and time control take the full height of
the row they are on.  Format the time spinner to center better and
remove the gap to give a clear indication it is attached to the time
control.  Hopefully this look sfine on macOS and Linux as well.