Search chain and replace plugin (#115)
* Some initial work on search sequences/chains + unique search id

* Flesh out edit search chain dialog

* Cleanup

- Simplify editor dialog setup with gridbagsizer
- Switch to listctrl (wxFormBuilder is freezing up with ListBox on OS X)
- Clean up some deprecations.

* Localize edit chain dialog

* Dialog to manage saved chains (incomplete)

* Finish up chain management

* Go back to ListBox with a workaround to fix tabbing

* Actual ability to chain multiple saved regex patterns

* Update tests

* Fix lint

* Localize search chain dialog

* Show literal status in save search dialog

* Perform validation on chains before search

Chains should check if all chained searches exist, and that the search
compiles.  Also clean up related code and make it more compact.  We
were also always setting regex version flags which doesn’t make much if
regex is disabled.

* Make chain wxCombo a wxChoice and remove divider

* When chains is enabled, change default focus

* Update comment

* Lint

* Rumcore doesn't decode strings directly

Rumcore will no longer decode strings to simplify the internal code.
If someone is using the API, they should decode it to Unicode or pass
it as a byte string for binary.  Only files will be decoded/encoded on
input/output. Technically everything is included so a user could detect
encoding.  But as the API is not documented right now, I won’t worry
about it, but in the future, maybe we can expose an easy detect
function that does everything the that happens for files.

* Fix replace related issues

* Revert: ensure we always return 1 record

Forgot that I require the API to always return 1 file record (or buffer
record in the case of a buffer replace).  Revert earlier change. and
ensure all loops abort.

* Fix search abort not working

* Update changelog

* Update dictionary with spelling

* Replace search button text when aborting

* Adjust putty patterns to work locally

* Git putty to work locally and on Travis

* Use existing "search for" object for chains

* Add replace plugin support + fixes

Add replace plugin support and fix issue where drop down list didn’t
hide the limit panel.  Also don’t display replace message with aborting
replace.  Also run flake8 in tox differently so we can avoid weird
flake8-putty patterns.

* Add .codecov.yml

* Don't let plugin refs get garbage collected

We had to pass the whole module in to keep plugin function working
because the module was getting garbage collected. We don’t want to
store the plugin in sys.module, so track in a local structure and
delete when we are done with it.

* Don't resize panel on search

* Fix path check

* Add format to settings and history file

* Visual tweaks

- Adjust layouts after chainging label names
- Don't adjust height when only graying out controls
- Disable and enable "replace plugin" directory picker when applicable.

* Handle special cases

- Disable replace plugin checkbox and save search button when chain
mode is enabled.

- When loading search, run event for chain mode, regex search, and/or
replace plugin when toggling their respective checkbox manually.

* Setup argument to avoid redirect

* Add support for replace plugin in tester

* Literal and Unicode changes

Literal will allow the Unicode flag and will allow full case in Regex
V0.  Tester will now allow testing with literal to allow pairing
literal with replace plugin.

* Patch gui and ensure literal flag is handled in tester

* Add some missing localization

* Update some documentation

- Add recent changes to changelog
- Add missing statement about POSIX flag

* Fix GUI inconsistencies

Visual inconsistencies in regard to text box size relative to siblings,
alignment of labels, etc. Also, content text box in tester dialog will
now allow entering tab characters on all platforms instead of navigating
to next control.

* Tweak date and time control

On each OS, the date and time object is a little different with how it
fits.  Try and ensure the date and time control take the full height of
the row they are on.  Format the time spinner to center better and
remove the gap to give a clear indication it is attached to the time
control.  Hopefully this look sfine on macOS and Linux as well.

* Tweak time ctrl

* Don't expand spinner (distorts on Ubuntu)

* Fix deprecation in messages lib

* Update saved project

* Some doc updates

* Fix spelling failure

* Add optional hint to add search diag

Add a hint to say comment is optional opposed to having it in the
label.  Also tweak some of the add search labels to be shorter.

* Rumcore tweaks

Denote RummageFileContent as a protected class.
Move Search object with Replace plugin object.
Add on_init method to Replaceplugin object.

* Update docs with more info on new features

* Don't list regex as optional (just optional to use)

* Fix spelling

* Update docs

New images, some re-wording and update theme CSS.

* Bump rev