Rework localization
Grouping strings together in file header isn’t good as the strings are
translated on import and miss the settings getting initialized.  Change
it so that each class to translate strings on instantiation. Or in
special cases does it on the fly.  This way we can reuse similar
strings as variables and see all the strings in the dialog, and have it
work correctly.