Ensure sane main window size
Usually, with English localization, the main window is only a little
bigger than 800 x 600.  But with localization this could be bigger if
strings are particularly long for a certain localization.  Screens are
really big these days, but just as a sanity check, we want to ensure
initial resizing and subsequent height resizing never make the window
bigger than the client window size.  The main window isn’t scrollable,
and we may never hit this issue, but if we ever do, maybe in the future
we can make the main window scrollable on overflow.