Handle fullscreen corner case
Detect (at least on mac) full screen corner cases.  On mac, during
fullscreen, the client area is sized differently as things like the
status bar are hidden.  So we track client size, and on resize, we see
if client size doesn’t add up to what we thought it should.  If the
size changed, we will optimize minimum size as we may have set minimum
size with incorrect assumptions.