Revert ensuring dialog is never bigger than window
I’m flip flopping on this.  I think a window being too big that can be
dragged is better than a window too small that can’t be scrolled and
loses functionality. I think if I want to go this route, I need to
convert the window into a scrollable window.