Move to flake8 3.0 (#128)
Ditch flake8-putty which is locked to <2.  It is mainly used to ignore naming issues with wxpython, but I don't have an issue with accidentally messing up naming conventions, and I can easily catch those kinds of issues in pull requests via a review.  Other issues were fixes that we ignored due to lazyness. Also, now we will run flake8 on Python3 instead of 2.  I want to prepare for when PY2 is no longer supported in 2020 as we will most likely drop PY2 support then ourselves.