Prevent rummage from loading local libs
Rummage can potentially load local libraries due to the way Python adds the current directory at the start of the path.  We remove this to prevent such things as rummage may be run in Python projects.  The only thing we can't prevent is the initial import of sys, but this is much safer now.