Binary detection tweaks
Use `string in otherstring` instead of regex to find null bytes as it is much 2x faster.  Chardet doesn't detect utf16 or utf32 without a bom, and since we don't have logic to general do this either, we can assume null means binary (some exceptions in XML). Detect 32bit middle endian encodings and fallback to binary as it is doubtful that a Python install supports this encoding (I doubt anyone actually uses it either). Don't use regex to detect bom to increase speed.