WINDOWFRAME color instead of BTNFACE
By default some sub panels get the color white which is ugly and unexpected in dialogs at time (maybe mainly macOS).  Parent panels don't seem to get this, so in the past, I set sub panels to BTNFACE as it had the expected color.  But it is probably more appropriate to use WINDOWFRAME. Go ahead and set all panels to WINDOWFRAME for consistency.  Will have to compare across platforms.