Internal HTML handling fixes and tweaks
`about:blank` links occur in HTML string pages, but if you link to some internal page, the Windows back end will try `about:<some path>`.  When we are executing HTML string pages, we don't want acccess to internal pages (at least not at this point in time).

For now, use base 64 encoding for images when using HTML strings. PathConverter's output is a little wrong, but it isn't entirely wrong.  Just not sure if we want to chagne it just to make this work on Windows.