New about dialog and document table of contents (#285)
* New about dialog

Ref #284

* Lint, docs, and locale

Fix lint update, changelog in docs, and regenerate locale catalog.

* Update internal docs

* Let web backend handle about:blank

* Internal HTML handling fixes and tweaks

`about:blank` links occur in HTML string pages, but if you link to some internal page, the Windows back end will try `about:<some path>`.  When we are executing HTML string pages, we don't want acccess to internal pages (at least not at this point in time).

For now, use base 64 encoding for images when using HTML strings. PathConverter's output is a little wrong, but it isn't entirely wrong.  Just not sure if we want to chagne it just to make this work on Windows.

* Restrict Linux and mac links to HTML string pages

Restrict HTML string to only inter page links and external URL, not other internal pages.

* Add toc to document pages

* Better table of contents themeing

* Initialize baseurl on Linux