No need to "kill" DirWalker
We don't need to "kill" DirWalker. We can use a simple break.
Originally, before we split out the logic into wcmatch, we were using
this class differently. There really is no reason to kill it as it
shouldn't have anything we want to process when we see the abort.
Additionally, in the future, I will remove the "kill" nonsense from the
WcMatch class.