Collapsible options (#319)
* Add collapsible controls

* Remember pane states and set localization strings for panes

* Update to a dev version and add changelog notes

* Fix spelling

* Generic collapsible control
This will allow us a consistent look across OS, and hopefully avoid
OS specific quirks with native controls.

* Simplify code

* Small partial workaround for mac

* Clean up current mac workaround and ensure works in Windows.

* Good tab traversal for mac and windows

* Small GUI sizer fix

* Small fix for autocomplete popup indexing

* Fix Linux tab issue

* Update internal docs

* Add support for tinting graphical assets

We can now tint image assets if needed. This will be useful for one
day when we properly support dark themes in macOS etc. For now, we
simply use it for collapsible pane buttons.

* Fix spelling errors