Do initial resize earlier on macOS and Windows
Linux (or maybe just GTK) has an issue where we have to resize after we
are loaded in on_loaded. We were doing this for all platforms, but this
causes an annoying visual side effect on other systems, so resize before
on loaded if not on Linux. Also, only do the additional focus for macOS,
as it is a workaround specifically targeted for macOS.