Resizing fixes related to collapsible panels
When the user switches to using re or regex in the preferences dialog,
we hide or show toggles and then force a resize. But with the addition
of the collapsible panels, are current methodology is not sufficient.
Update logic to properly resize in this scenario. Also, cleanup initial
on load workaround related to sizing and focusing.