Linux tab interface color fix (#347)
Linux should not use the page background to paint the tab control and
collapsible controls, but instead should use the tab control. Colors are
not always reported exactly as they actually appear, so we will acquire
the default panel color in Linux and just use it for the tabbed
interface.