Consistently deal with colors in wxPython
Before we were dealing with colors as RGB (0-255) or packed wxPython
abgr. This would make things a little confusing coming back.

Moving forward, always deal with colors as packed abgr. We can always
get this value. This simplifies things and we don't have to remember
cases where we have to use one vs the other.

Raw image values is the one exception, but we'll manually convert
between the 0 - 255 and 0 - 1 range.